feat: handle multiplication with float values

+++ matmul.c
@@ -1,6 +1,6 @@
 #include <stdlib.h>
 
-void matrix_multiply_ijk(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_ijk(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int i = 0; i < A_row_size; i++) {
         for (int j = 0; j < B_col_size; j++) {
             for (int k = 0; k < B_row_size; k++) {
@@ -10,7 +10,7 @@ void matrix_multiply_ijk(int **A, int **B, int **C, int A_row_size, int B_col_si
     }
 }
 
-void matrix_multiply_ikj(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_ikj(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int i = 0; i < A_row_size; i++) {
         for (int k = 0; k < B_row_size; k++) {
             for (int j = 0; j < B_col_size; j++) {
@@ -20,7 +20,7 @@ void matrix_multiply_ikj(int **A, int **B, int **C, int A_row_size, int B_col_si
     }
 }
 
-void matrix_multiply_jik(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_jik(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int j = 0; j < B_col_size; j++) {
         for (int i = 0; i < A_row_size; i++) {
             for (int k = 0; k < B_row_size; k++) {
@@ -30,7 +30,7 @@ void matrix_multiply_jik(int **A, int **B, int **C, int A_row_size, int B_col_si
     }
 }
 
-void matrix_multiply_jki(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_jki(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int j = 0; j < B_col_size; j++) {
         for (int k = 0; k < B_row_size; k++) {
             for (int i = 0; i < A_row_size; i++) {
@@ -40,7 +40,7 @@ void matrix_multiply_jki(int **A, int **B, int **C, int A_row_size, int B_col_si
     }
 }
 
-void matrix_multiply_kij(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_kij(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int k = 0; k < B_row_size; k++) {
         for (int i = 0; i < A_row_size; i++) {
             for (int j = 0; j < B_col_size; j++) {
@@ -50,7 +50,7 @@ void matrix_multiply_kij(int **A, int **B, int **C, int A_row_size, int B_col_si
     }
 }
 
-void matrix_multiply_kji(int **A, int **B, int **C, int A_row_size, int B_col_size, int B_row_size) {
+void matrix_multiply_kji(float **A, float **B, float **C, int A_row_size, int B_col_size, int B_row_size) {
     for (int k = 0; k < B_row_size; k++) {
         for (int j = 0; j < B_col_size; j++) {
             for (int i = 0; i < A_row_size; i++) {
